Calculate strike force from ball speed
This isn't working quite right yet. The ball seems to be going over
twice as fast as expected when using the dirve (and I haven't tested the
other clubs).

Also, I can't set the correct mass in the on the gameObject and have to
rely on a constant in the script. If I do set the gameObject mass to the
proper 46 grams of a golf ball, the force applied absolutely launches
the ball. So for now I have set it to 1kilogram.